+++ scripts/joint_torque_springs.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (c) 2013-2014, Rethink Robotics
+# Copyright (c) 2013-2015, Rethink Robotics
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -77,7 +77,7 @@ class JointSprings(object):
 
         # create cuff disable publisher
         cuff_ns = 'robot/limb/' + limb + '/suppress_cuff_interaction'
-        self._pub_cuff_disable = rospy.Publisher(cuff_ns, Empty)
+        self._pub_cuff_disable = rospy.Publisher(cuff_ns, Empty, queue_size=1)
 
         # verify robot is enabled
         print("Getting robot state... ")