+++ scripts/joint_trajectory_client.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 
-# Copyright (c) 2013-2014, Rethink Robotics
+# Copyright (c) 2013-2015, Rethink Robotics
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -60,6 +60,8 @@ class Trajectory(object):
             FollowJointTrajectoryAction,
         )
         self._goal = FollowJointTrajectoryGoal()
+        self._goal_time_tolerance = rospy.Time(0.1)
+        self._goal.goal_time_tolerance = self._goal_time_tolerance
         server_up = self._client.wait_for_server(timeout=rospy.Duration(10.0))
         if not server_up:
             rospy.logerr("Timed out waiting for Joint Trajectory"
@@ -90,6 +92,7 @@ class Trajectory(object):
 
     def clear(self, limb):
         self._goal = FollowJointTrajectoryGoal()
+        self._goal.goal_time_tolerance = self._goal_time_tolerance
         self._goal.trajectory.joint_names = [limb + '_' + joint for joint in \
             ['s0', 's1', 'e0', 'e1', 'w0', 'w1', 'w2']]
 
@@ -131,6 +134,10 @@ def main():
 
     traj = Trajectory(limb)
     rospy.on_shutdown(traj.stop)
+    # Command Current Joint Positions first
+    limb_interface = baxter_interface.limb.Limb(limb)
+    current_angles = [limb_interface.joint_angle(joint) for joint in limb_interface.joint_names()]
+    traj.add_point(current_angles, 0.0)
 
     p1 = positions[limb]
     traj.add_point(p1, 7.0)